+++ CHANGELOG.md
@@ -4,6 +4,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## 0.0.2 - 2017-07-21
 ### Added
 - Include both the CSS var setters and getters to the outputted map
+- A test to confirm the outputted map file is correct
+
+### Changed
+- Use a more flexible prefix and suffix option instead of a var name 
 
 ## 0.0.1 - 2017-07-12
 ### Added